app/requirements.txt
fastapi
uvicorn
pydantic==1.10.11
python-dotenv
debugpy
aiohttp
motor